[Refactor] #66 구조 동물 API 활용 로직을 서비스 레이어로 분리
- ProtectingAnimalApiClient 의 코드를 최대한 순수하게 유지
- ProtectingAnimalApiClient 내부에 존재하던 비즈니스 로직을 서비스 레이어로 분리
