+++ src/main/java/com/kuit/findyou/domain/report/service/sync/ProtectingReportSyncService.java
@@ -1,0 +1,6 @@
+package com.kuit.findyou.domain.report.service.sync;
+
+public interface ProtectingReportSyncService {
+
+    void syncProtectingReports();
+}
